new file:   .gitignore 	modified:   main/bmp390.c 	modified:   main/fc_comm.c 	modified:   main/main.c 	modified:   main/typedefs.h

+++ main/typedefs.h
@@ -39,15 +39,17 @@ typedef struct
   uint8_t waypoint_mission_status;
 } flight_t;
 
+
 typedef struct
 {
-  int16_t pitch;
-  int16_t roll;
-  uint16_t heading;
-  int16_t pitch_dps;
-  int16_t roll_dps;
-  int16_t yaw_dps;
+  float pitch;
+  float roll;
+  float heading;
+  float pitch_dps;
+  float roll_dps;
+  float yaw_dps;
 } data_1_t;
+
 typedef struct
 {
   int16_t velocity_x_ms;
